delete answer use case

+++ src/domain/forum/application/use-cases/delete-answer.ts
@@ -11,6 +11,7 @@ export class DeleteAnswerUseCase {
   constructor(private answersRepository: AnswerRepository) {}
 
   async execute({
+    authorId,
     answerId,
   }: DeleteAnswerUseCaseRequest): Promise<DeleteAnswerUseCaseResponse> {
     const answer = await this.answersRepository.findById(answerId)
@@ -19,10 +20,12 @@ export class DeleteAnswerUseCase {
       throw new Error('Answer not found.')
     }
 
+    if (authorId !== answer.authorId.toString()) {
+      throw new Error('Not allowed.')
+    }
+
     await this.answersRepository.delete(answer)
 
-    return {
-      answer,
-    }
+    return {}
   }
 }
